move common structs/enums to common folder
