+++ esdt-safe/Cargo.toml
@@ -14,9 +14,6 @@ wasm-output-mode = [ "elrond-wasm-node",]
 [dependencies.transaction]
 path = "../common/transaction"
 
-[dependencies.ethereum-fee-prepay]
-path = "../ethereum-fee-prepay"
-
 [dependencies.elrond-wasm]
 version = "0.14.1"
